# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/financial": {
      "post": {
        "operationId": "FinancialController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Financial criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Criar Financial",
        "tags": [
          "Financial"
        ]
      },
      "get": {
        "operationId": "FinancialController_findAll",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Listar todos Financial",
        "tags": [
          "Financial"
        ]
      }
    },
    "/financial/summary": {
      "get": {
        "operationId": "FinancialController_getSummary",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation getSummary",
        "tags": [
          "Financial"
        ]
      }
    },
    "/financial/dre": {
      "get": {
        "operationId": "FinancialController_getDre",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation getDre",
        "tags": [
          "Financial"
        ]
      }
    },
    "/financial/projection": {
      "get": {
        "operationId": "FinancialController_getProjection",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation getProjection",
        "tags": [
          "Financial"
        ]
      }
    },
    "/financial/{id}": {
      "get": {
        "operationId": "FinancialController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Buscar um Financial",
        "tags": [
          "Financial"
        ]
      },
      "put": {
        "operationId": "FinancialController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Atualizar Financial",
        "tags": [
          "Financial"
        ]
      },
      "delete": {
        "operationId": "FinancialController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operação realizada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Remover Financial",
        "tags": [
          "Financial"
        ]
      }
    },
    "/financial/{id}/generate-pix": {
      "post": {
        "operationId": "FinancialController_generatePix",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Financial criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation generatePix",
        "tags": [
          "Financial"
        ]
      }
    },
    "/financial/webhook/mercadopago": {
      "post": {
        "operationId": "FinancialController_handleWebhook",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Financial criado com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Não autorizado."
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Operation handleWebhook",
        "tags": [
          "Financial"
        ]
      }
    },
...
  },
  "info": {
    "title": "Click Marido API",
    "description": "API documentation for Click Marido ERP/CRM",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {
          "companyId": {
            "type": "string",
            "description": "Campo companyId",
            "example": "exemplo"
          },
          "type": {
            "type": "string",
            "description": "Campo type",
            "example": "exemplo"
          },
          "category": {
            "type": "string",
            "description": "Campo category",
            "example": "exemplo"
          },
          "value": {
            "type": "number",
            "description": "Campo value",
            "example": 1
          },
          "description": {
            "type": "string",
            "description": "Campo description",
            "example": "exemplo"
          },
          "transactionDate": {
            "type": "string",
            "description": "Campo transactionDate",
            "example": "exemplo"
          },
          "dueDate": {
            "type": "string",
            "description": "Campo dueDate",
            "example": "exemplo"
          },
          "status": {
            "type": "string",
            "description": "Campo status",
            "example": "exemplo"
          },
          "paidAt": {
            "type": "string",
            "description": "Campo paidAt",
            "example": "exemplo"
          }
        },
        "required": [
          "companyId",
          "type",
          "category",
          "value",
          "transactionDate"
        ]
      },
      "UpdateTransactionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Campo type",
            "example": "exemplo"
          },
          "category": {
            "type": "string",
            "description": "Campo category",
            "example": "exemplo"
          },
          "value": {
            "type": "number",
            "description": "Campo value",
            "example": 1
          },
          "description": {
            "type": "string",
            "description": "Campo description",
            "example": "exemplo"
          },
          "transactionDate": {
            "type": "string",
            "description": "Campo transactionDate",
            "example": "exemplo"
          },
          "dueDate": {
            "type": "string",
            "description": "Campo dueDate",
            "example": "exemplo"
          },
          "status": {
            "type": "string",
            "description": "Campo status",
            "example": "exemplo"
          },
          "paidAt": {
            "type": "string",
            "description": "Campo paidAt",
            "example": "exemplo"
          }
        }
      },
...
    }
  }
}